more code moving from noddy to pigfarm
