+++ karmapi/noddy.py
@@ -37,21 +37,6 @@ class Magic(pigfarm.MagicCarpet):
         if data:
             self.process_data()
 
-        self.add_event_map(' ', self.next_group)
-
-
-    async def next_group(self):
-        """ Next group """
-        self.group += 1
-
-        if self.group == len(self.groups):
-            self.group = 0
-
-        await self.event.put(self.group)
-
-
- 
-
     async def tester(self, sleep=2):
 
         print('NUMBER OF TESTS', self.tests)
@@ -79,39 +64,6 @@ class Magic(pigfarm.MagicCarpet):
         
         await pigfarm.sleep(sleep)
 
-
-    def draw_plot(self):
-
-        group = self.groups[self.group]
-        frame = self.frames[group]
-
-        axes = self.subplots[0]
-        axes.clear()
-
-        # sort columns on mean
-        mean = frame.mean()
-        mean.sort_values(inplace=True)
-        frame = frame.ix[:, mean.index]
-        
-        col_colours = []
-        for label in frame.columns:
-            data = frame[label].copy()
-            data.sort_values(inplace=True)
-            if self.log:
-                patch = axes.semilogy(data.values, label=label)
-            else:
-                patch = axes.plot(data.values, label=label)
-
-            col_colours.append(patch[0].get_color())
-
-        from matplotlib import colors
-        col_colours = [colors.to_rgba(x, 0.2) for x in col_colours]
-
-        self.axes = self.subplots[1]
-        self.axes.clear()
-        self.draw_table(frame, loc='center', title=group, col_colours=col_colours)
-        self.draw()
-        
     async def run(self):
 
         await pigfarm.spawn(self.load_data())
